Make the repo compilable on MacOS with Apple Clang

+++ t86/dbg-cli/CLI.h
@@ -148,8 +148,8 @@ public:
         if (address >= text_size) {
             return;
         }
-        size_t begin = static_cast<unsigned>(range) > address ? 0 : address - range;
-        size_t end = std::min(text_size, address + range + 1);
+        size_t begin = static_cast<size_t>(range) > address ? 0 : address - range;
+        size_t end = std::min(text_size, static_cast<size_t>(address) + range + 1);
         auto inst = process.ReadText(begin, end - begin);
         for (size_t i = 0; i < inst.size(); ++i) {
             uint64_t curr_addr = i + begin;
